requirements.txt
crewai>=0.30.0
litellm>=1.0.0
nltk>=3.8.1
python-dotenv>=1.0.1
pydantic>=2.0
fpdf>=1.7.2